docs: BaseApiTest 추가
- 공통 설정 클래스 `BaseApiTest` 추가로 중복 코드 제거
- `ProjectApiTest`와 `ArticleApiTest`가 `BaseApiTest`를 상속하도록 수정
- RestDocs 문서화 코드 가독성과 재사용성 개선
- 불필요한 필터 설정 제거 및 `document` 메서드 활용으로 코드 간결화
